添加同步数据处理方法,删除方法...

+++ app/src/main/java/demo/ringares/com/ical4jdemo/bean/LocationDataBean.java
@@ -8,8 +8,8 @@ package demo.ringares.com.ical4jdemo.bean;
 public class LocationDataBean {
     public int location_id;//主键 本地id
     public int location_event_id;// 对应event_id
-    public double location_lat;// 地址纬度
-    public double location_lon;// 地址经度
+    public float location_lat;// 地址纬度
+    public float location_lon;// 地址经度
     public String location_city;// 地址城市
     public String location_country;// 地址国家
     public String location_desc;// 地址整体详细描述
